refactor(#320): Capsulated the grel-string function into a separate structure. Put the function decorator into a separate file to hinder circular imports and for better separation of concerns.
